Add CI and stuff (#1)
* Add .gitignore

* Add CI and build files

* Add __init__.py

* Format app with `black`

* Update README

* Comment out test step

* Fix `mypy` errors

* Format app with `black`

* Fix `mypy` errors

* Update README

* Update README

+++ classification/features.py
@@ -13,7 +13,7 @@ def contrast_flower(image: np.ndarray) -> float:
         float: Standard deviation of brightness values.
     """
     gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
-    return np.std(gray_image)
+    return float(np.std(gray_image))  # type: ignore[arg-type]
 
 
 def amount_of_green(image: np.ndarray) -> int:
@@ -83,7 +83,7 @@ def avg_saturation(image: np.ndarray) -> float:
     """
     hsv_image = cv2.cvtColor(image, cv2.COLOR_BGR2HSV)
     saturation = hsv_image[:, :, 1]  # Extract the saturation channel
-    return float(np.mean(saturation))
+    return float(np.mean(saturation))  # type: ignore[arg-type]
 
 
 def avg_brightness(image: np.ndarray) -> float:
@@ -98,7 +98,7 @@ def avg_brightness(image: np.ndarray) -> float:
     """
     hsv_image = cv2.cvtColor(image, cv2.COLOR_BGR2HSV)
     brightness = hsv_image[:, :, 2]  # Extract the brightness channel
-    return float(np.mean(brightness))
+    return float(np.mean(brightness))  # type: ignore[arg-type]
 
 
 def circularity(image: np.ndarray) -> float:
@@ -121,7 +121,7 @@ def circularity(image: np.ndarray) -> float:
     perimeter = cv2.arcLength(contour, True)
     if perimeter == 0:
         return 0.0
-    return 4 * np.pi * area / (perimeter ** 2)
+    return 4 * np.pi * area / (perimeter**2)
 
 
 def bud_shape(image: np.ndarray) -> float:
@@ -135,8 +135,12 @@ def bud_shape(image: np.ndarray) -> float:
         float: The elongation ratio (higher values for elongated shapes).
     """
     gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
-    _, binary_image = cv2.threshold(gray_image, 50, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU)
-    contours, _ = cv2.findContours(binary_image, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
+    _, binary_image = cv2.threshold(
+        gray_image, 50, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU
+    )
+    contours, _ = cv2.findContours(
+        binary_image, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE
+    )
 
     if not contours:
         return 0.0
@@ -160,8 +164,12 @@ def count_objects(image: np.ndarray) -> int:
     """
     gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
     blurred_image = cv2.GaussianBlur(gray_image, (5, 5), 0)
-    _, binary_image = cv2.threshold(blurred_image, 50, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU)
-    contours, _ = cv2.findContours(binary_image, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
+    _, binary_image = cv2.threshold(
+        blurred_image, 50, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU
+    )
+    contours, _ = cv2.findContours(
+        binary_image, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE
+    )
     return sum(1 for contour in contours if cv2.contourArea(contour) > 500)
 
 
@@ -176,13 +184,18 @@ def contour_complexity(image: np.ndarray) -> float:
         float: Average perimeter-to-area ratio of contours.
     """
     gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
-    _, binary_image = cv2.threshold(gray_image, 50, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU)
-    contours, _ = cv2.findContours(binary_image, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
+    _, binary_image = cv2.threshold(
+        gray_image, 50, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU
+    )
+    contours, _ = cv2.findContours(
+        binary_image, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE
+    )
     complexities = [
         cv2.arcLength(contour, True) / (cv2.contourArea(contour) + 1e-5)
-        for contour in contours if cv2.contourArea(contour) > 500
+        for contour in contours
+        if cv2.contourArea(contour) > 500
     ]
-    return np.mean(complexities) if complexities else 0.0
+    return float(np.mean(complexities)) if complexities else 0.0
 
 
 def object_density(image: np.ndarray) -> float:
@@ -197,7 +210,9 @@ def object_density(image: np.ndarray) -> float:
     """
     gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
     _, binary_image = cv2.threshold(gray_image, 200, 255, cv2.THRESH_BINARY)
-    contours, _ = cv2.findContours(binary_image, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
+    contours, _ = cv2.findContours(
+        binary_image, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE
+    )
     total_area = image.shape[0] * image.shape[1]
     object_count = sum(1 for contour in contours if cv2.contourArea(contour) > 300)
     return object_count / (total_area / 10000)
@@ -214,8 +229,12 @@ def elongated_shapes_ratio(image: np.ndarray) -> float:
         float: The ratio of elongated objects (ratio > 3 for elongated shapes).
     """
     gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
-    _, binary_image = cv2.threshold(gray_image, 50, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU)
-    contours, _ = cv2.findContours(binary_image, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
+    _, binary_image = cv2.threshold(
+        gray_image, 50, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU
+    )
+    contours, _ = cv2.findContours(
+        binary_image, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE
+    )
     elongated_count = 0
     total_count = 0
